requirements.txt
Django==1.6.2
MySQL-python==1.2.5
django-auth-ldap==1.2.0
django-filter==0.7
djangorestframework==2.3.13
python-ldap==2.4.15
requests==2.2.1
wsgiref==0.1.2
